REV-52 Rename initialActions to messageActions.
